소셜 로그인 로직 수정 및 user 파일 분리
