+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/mail/persistence/EmailVerificationRepository.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.toyproject.memoWithTags.user.persistence
+package com.wafflestudio.toyproject.memoWithTags.mail.persistence
 
 import org.springframework.data.jpa.repository.JpaRepository
 import java.time.LocalDateTime
