+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/mail/service/NoOpMailService.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.toyproject.memoWithTags.user.service
+package com.wafflestudio.toyproject.memoWithTags.mail.service
 
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
