+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/social_login/dto/GoogleUser.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.toyproject.memoWithTags.user.dto
+package com.wafflestudio.toyproject.memoWithTags.social_login.dto
 
 data class GoogleOAuthToken(
     val access_token: String,
