+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/user/GoogleUtil.kt
@@ -1,8 +1,8 @@
 package com.wafflestudio.toyproject.memoWithTags.user
 
 import com.wafflestudio.toyproject.memoWithTags.exception.OAuthRequestException
-import com.wafflestudio.toyproject.memoWithTags.user.dto.GoogleOAuthToken
-import com.wafflestudio.toyproject.memoWithTags.user.dto.GoogleProfile
+import com.wafflestudio.toyproject.memoWithTags.social_login.dto.GoogleOAuthToken
+import com.wafflestudio.toyproject.memoWithTags.social_login.dto.GoogleProfile
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpEntity
