+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/user/KakaoUtil.kt
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.core.JsonProcessingException
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import com.wafflestudio.toyproject.memoWithTags.exception.OAuthRequestException
-import com.wafflestudio.toyproject.memoWithTags.user.dto.KakaoOAuthToken
-import com.wafflestudio.toyproject.memoWithTags.user.dto.KakaoProfile
+import com.wafflestudio.toyproject.memoWithTags.social_login.dto.KakaoOAuthToken
+import com.wafflestudio.toyproject.memoWithTags.social_login.dto.KakaoProfile
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpEntity
