--- src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/user/service/MailService.kt
@@ -1,5 +1,0 @@
-package com.wafflestudio.toyproject.memoWithTags.user.service
-
-interface MailService {
-    fun sendMail(toEmail: String, title: String, content: String)
-}
