+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/user/service/UserService.kt
@@ -8,11 +8,12 @@ import com.wafflestudio.toyproject.memoWithTags.exception.MailVerificationExcept
 import com.wafflestudio.toyproject.memoWithTags.exception.SignInInvalidException
 import com.wafflestudio.toyproject.memoWithTags.exception.UserNotFoundException
 import com.wafflestudio.toyproject.memoWithTags.user.JwtUtil
-import com.wafflestudio.toyproject.memoWithTags.user.controller.EmailVerification
+import com.wafflestudio.toyproject.memoWithTags.mail.EmailVerification
 import com.wafflestudio.toyproject.memoWithTags.user.controller.User
 import com.wafflestudio.toyproject.memoWithTags.user.dto.UserResponse.RefreshTokenResponse
-import com.wafflestudio.toyproject.memoWithTags.user.persistence.EmailVerificationEntity
-import com.wafflestudio.toyproject.memoWithTags.user.persistence.EmailVerificationRepository
+import com.wafflestudio.toyproject.memoWithTags.mail.persistence.EmailVerificationEntity
+import com.wafflestudio.toyproject.memoWithTags.mail.persistence.EmailVerificationRepository
+import com.wafflestudio.toyproject.memoWithTags.mail.service.MailService
 import com.wafflestudio.toyproject.memoWithTags.user.persistence.UserEntity
 import com.wafflestudio.toyproject.memoWithTags.user.persistence.UserRepository
 import org.mindrot.jbcrypt.BCrypt
@@ -31,6 +32,9 @@ class UserService(
 ) {
     private val logger = LoggerFactory.getLogger(UserService::class.java)
 
+    /**
+     * 자체 로그인 과정 중 회원가입을 구현한 함수
+     */
     @Transactional
     fun register(
         email: String,
@@ -38,6 +42,7 @@ class UserService(
     ): User {
         if (userRepository.findByEmail(email) != null) throw EmailAlreadyExistsException()
         val encryptedPassword = BCrypt.hashpw(password, BCrypt.gensalt())
+        // 메일 인증이 이루어지기 전까지 User의 verified 필드는 false이다.
         val userEntity = userRepository.save(
             UserEntity(
                 email = email,
@@ -49,6 +54,9 @@ class UserService(
         return User.fromEntity(userEntity)
     }
 
+    /**
+     * 회원가입 또는 비밀번호 변경 요청 후 인증용 메일을 발송하는 함수
+     */
     fun sendCodeToEmail(
         email: String
     ) {
@@ -72,6 +80,9 @@ class UserService(
         }
     }
 
+    /**
+     * 인증 메일에 포함될 인증 코드를 랜덤으로 생성하는 함수. 6자리 숫자를 생성한다.
+     */
     private fun createVerificationCode(email: String): EmailVerification {
         val randomCode: String = (100000..999999).random().toString()
         val codeEntity = EmailVerificationEntity(
@@ -82,6 +93,9 @@ class UserService(
         return EmailVerification.fromEntity(emailVerificationRepository.save(codeEntity))
     }
 
+    /**
+     * 메일로 보내진 인증 번호와 유저가 입력한 인증 번호가 일치하는지 검증하는 함수
+     */
     @Transactional
     fun verifyEmail(
         email: String,
@@ -90,10 +104,14 @@ class UserService(
         val verification = emailVerificationRepository.findByEmailAndCode(email, code) ?: throw MailVerificationException()
         if (verification.expiryTime.isBefore(LocalDateTime.now())) throw AuthenticationFailedException()
         val userEntity = userRepository.findByEmail(verification.email)
+        // 인증 성공 시, user의 Verified 필드가 true로 바뀌어 정식 회원이 된다.
         userEntity!!.verified = true
         return true
     }
 
+    /**
+     * 회원가입된 유저의 자체 로그인 로직을 수행하는 함수
+     */
     @Transactional
     fun login(
         email: String,
@@ -109,6 +127,9 @@ class UserService(
         )
     }
 
+    /**
+     * 비밀번호 변경을 위해 보내진 메일 인증을 완료하고, 비밀번호를 변경하는 함수
+     */
     @Transactional
     fun resetPassword(
         email: String,
@@ -121,6 +142,9 @@ class UserService(
         }
     }
 
+    /**
+     * 유저 토큰을 받아 유저 정보를 반환하는 함수
+     */
     @Transactional
     fun authenticate(
         accessToken: String
@@ -132,6 +156,9 @@ class UserService(
         return User.fromEntity(userEntity)
     }
 
+    /**
+     * accessToken 만료 시 refreshToken을 통해 유저를 확인하고 새로운 accessToken을 발급하는 함수
+     */
     fun refreshToken(refreshToken: String): RefreshTokenResponse {
         if (!JwtUtil.isValidToken(refreshToken)) {
             throw InValidTokenException()
@@ -150,17 +177,26 @@ class UserService(
         )
     }
 
+    /**
+     * 해당 메일의 User를 찾는 함수. 없으면 예외를 발생시킨다.
+     */
     @Transactional
     fun getUserEntityByEmail(email: String): UserEntity {
         return userRepository.findByEmail(email) ?: throw UserNotFoundException()
     }
 
+    /**
+     * 메일 정오에 만료된 인증 코드 엔티티를 삭제하는 함수
+     */
     @Transactional
     @Scheduled(cron = "0 0 12 * * ?") // 매일 정오에 만료 코드 삭제
     fun deleteExpiredVerificationCode() {
         emailVerificationRepository.deleteByExpiryTimeBefore(LocalDateTime.now())
     }
 
+    /**
+     * 매일 정오에 메일 인증이 되지 않은 유저를 삭제하는 함수
+     */
     @Transactional
     @Scheduled(cron = "0 0 12 * * ?") // 매일 정오에 미인증 사용자 삭제
     fun deleteUnverifiedUser() {
